content/media/test/test_seek.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: seek tests</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
  <script type="text/javascript" src="seek1.js"></script>
  <script type="text/javascript" src="seek2.js"></script>
  <script type="text/javascript" src="seek3.js"></script>
  <script type="text/javascript" src="seek4.js"></script>
  <script type="text/javascript" src="seek5.js"></script>
  <script type="text/javascript" src="seek6.js"></script>
  <script type="text/javascript" src="seek7.js"></script>
  <script type="text/javascript" src="seek8.js"></script>
  <script type="text/javascript" src="seek9.js"></script>
  <script type="text/javascript" src="seek10.js"></script>
  <script type="text/javascript" src="seek11.js"></script>
  <script type="text/javascript" src="seek12.js"></script>
  <script type="text/javascript" src="seek13.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.requestLongerTimeout(3);
var manager = new MediaTestManager;

// https://bugzilla.mozilla.org/show_bug.cgi?id=634747
if (navigator.platform.startsWith("Win")) {
  SimpleTest.expectAssertions(0, 5);
} else {
  // This is "###!!! ASSERTION: Page read cursor should be inside range: 'mPageOffset <= endOffset'"
  // https://bugzilla.mozilla.org/show_bug.cgi?id=846769
  SimpleTest.expectAssertions(0, 5);
}

const NUM_SEEK_TESTS = 13;

function createTestArray() {
  var tests = [];
  var tmpVid = document.createElement("video");

  for (var testNum=0; testNum<gSeekTests.length; testNum++) {
    var test = gSeekTests[testNum];
    if (!tmpVid.canPlayType(test.type)) {
      continue;
    }

    for (var i = 1; i <= NUM_SEEK_TESTS; ++i) {
      var t = new Object;
      t.name = test.name;
      t.type = test.type;
      t.duration = test.duration;
      t.number = i;
      tests.push(t);
    }
  }
  return tests;
}

function startTest(test, token) {
  var v = document.createElement('video');
  v.token = token += "-seek" + test.number + ".js";
  manager.started(v.token);
  v.src = test.name;
  v.preload = "metadata";
  document.body.appendChild(v);
  var name = test.name + " seek test " + test.number;
  var localIs = function(name) { return function(a, b, msg) {
    is(a, b, name + ": " + msg);
  }}(name);
  var localOk = function(name) { return function(a, msg) {
    ok(a, name + ": " + msg);
  }}(name);
  var localFinish = function(v, manager) { return function() {
    v.onerror = null;
    removeNodeAndSource(v);
    dump("SEEK-TEST: Finished " + name + " token: " + v.token + "\n");
    manager.finished(v.token);
  }}(v, manager);
  dump("SEEK-TEST: Started " + name + "\n");
  window['test_seek' + test.number](v, test.duration/2, localIs, localOk, localFinish);
}

manager.runTests(createTestArray(), startTest);

</script>
</pre>
</body>
</html>
